+++ jkube-kit/jkube-kit-quarkus/src/test/java/org/eclipse/jkube/quarkus/generator/QuarkusGeneratorTest.java
@@ -29,7 +29,7 @@ import java.util.Properties;
 import org.eclipse.jkube.generator.api.GeneratorContext;
 import org.eclipse.jkube.kit.config.image.ImageConfiguration;
 import org.eclipse.jkube.kit.common.JavaProject;
-import org.eclipse.jkube.kit.config.image.build.OpenShiftBuildStrategy;
+import org.eclipse.jkube.kit.config.image.build.JKubeBuildStrategy;
 import org.eclipse.jkube.kit.config.resource.ProcessorConfig;
 import org.eclipse.jkube.kit.config.resource.RuntimeMode;
 import org.junit.Before;
@@ -174,7 +174,7 @@ public class QuarkusGeneratorTest {
             config.getConfig("test-generator", "from"); result = configFrom; minTimes = 0;
             config.getConfig("test-generator", "fromMode"); result = configFromMode; minTimes = 0;
             ctx.getRuntimeMode(); result = RuntimeMode.OPENSHIFT; minTimes = 0;
-            ctx.getStrategy(); result = OpenShiftBuildStrategy.s2i; minTimes = 0;
+            ctx.getStrategy(); result = JKubeBuildStrategy.s2i; minTimes = 0;
         }};
         // @formatter:on
     }